+++ src/ui/mod.rs
@@ -10,9 +10,8 @@ use self::help::draw_help;
 use self::overview::draw_overview;
 use self::utilization::draw_utilization;
 use self::utils::{
-  centered_rect, horizontal_chunks_with_margin, layout_block, style_default, style_failure,
-  style_help, style_main_background, style_primary, style_secondary, title_style_logo,
-  vertical_chunks,
+  horizontal_chunks_with_margin, layout_block, style_default, style_failure, style_help,
+  style_main_background, style_primary, style_secondary, title_style_logo, vertical_chunks,
 };
 use super::app::{App, RouteId};
 
@@ -29,30 +28,38 @@ static HIGHLIGHT: &str = "=> ";
 pub fn draw<B: Backend>(f: &mut Frame<B>, app: &mut App) {
   let block = Block::default().style(style_main_background(app.light_theme));
   f.render_widget(block, f.size());
-  let chunks = vertical_chunks(vec![Constraint::Length(3), Constraint::Min(0)], f.size());
+
+  let chunks = if !app.api_error.is_empty() {
+    let chunks = vertical_chunks(
+      vec![
+        Constraint::Length(3),
+        Constraint::Length(3),
+        Constraint::Min(0),
+      ],
+      f.size(),
+    );
+    draw_app_error(f, app, chunks[1]);
+    chunks
+  } else {
+    vertical_chunks(vec![Constraint::Length(3), Constraint::Min(0)], f.size())
+  };
 
   // draw header and logo
   draw_app_header(f, app, chunks[0]);
 
+  let last_chunk = chunks[chunks.len() - 1];
   match app.get_current_route().id {
     RouteId::HelpMenu => {
-      draw_help(f, app, chunks[1]);
-    }
-    RouteId::Error => {
-      if app.api_error.is_empty() {
-        draw_overview(f, app, chunks[1]);
-      } else {
-        draw_error_popup(f, app, chunks[1]);
-      }
+      draw_help(f, app, last_chunk);
     }
     RouteId::Contexts => {
-      draw_contexts(f, app, chunks[1]);
+      draw_contexts(f, app, last_chunk);
     }
     RouteId::Utilization => {
-      draw_utilization(f, app, chunks[1]);
+      draw_utilization(f, app, last_chunk);
     }
     _ => {
-      draw_overview(f, app, chunks[1]);
+      draw_overview(f, app, last_chunk);
     }
   }
 }
@@ -97,11 +104,11 @@ fn draw_header_text<B: Backend>(f: &mut Frame<B>, app: &mut App, area: Rect) {
   f.render_widget(paragraph, area);
 }
 
-fn draw_error_popup<B: Backend>(f: &mut Frame<B>, app: &mut App, size: Rect) {
+fn draw_app_error<B: Backend>(f: &mut Frame<B>, app: &mut App, size: Rect) {
   let block = Block::default()
     .title("Error | close <esc>")
+    .style(style_failure())
     .borders(Borders::ALL);
-  let area = centered_rect(60, 20, size);
 
   let mut text = Text::from(app.api_error.clone());
   text.patch_style(style_failure());
@@ -110,5 +117,5 @@ fn draw_error_popup<B: Backend>(f: &mut Frame<B>, app: &mut App, size: Rect) {
     .style(style_primary())
     .block(block)
     .wrap(Wrap { trim: true });
-  f.render_widget(paragraph, area);
+  f.render_widget(paragraph, size);
 }